Handle the restore bug and remove unnecessary info
Sometimes restoring fails. And, if it fails, it fails badly. It means
some windows persist minimized. Often it happens when you have about 6
or more splitted windows and there is a right column window (a window
with max height), eg. a Tagbar window. Perhaps in complex split window
patterns vim is renumerating windows while resing?

Anyway, the implemented workaround is to detect such case and perform
<C-w>= to give it a try to recover anything.