Only unmaximize when currently maximized.
I find the plugin useful, but it's hard for me to remember the state the plugin is in, especially if the maximization was done some time ago and I have modified the window layout in the meantime. I think it's better to extend the logic so that demaximization is only performed when the current window is still in maximized state. When other windows' sizes have been increased in the meantime, re-maximize instead of re-applying the stored, outdated window layout.